+++ main.py
@@ -51,14 +51,17 @@ def main(setTargetURL, setUserlist, setPasslist, setNumberThreads):
 			worker.daemon = True
 			worker.start()
 
-	except (KeyboardInterrupt, SystemExit):
-		utils.printf("Program has terminated!!!", "bad")
-		for worker in workers:
-			worker.join()
-		sys.exit(1)
+	#except (KeyboardInterrupt, SystemExit):
+	except KeyboardInterrupt:# as error:
+		# for worker in workers:
+		# 	worker.join()
+		utils.die("Terminated by user!", KeyboardInterrupt)
+		
+	except SystemExit:# as error
+		utils.die("Terminated by system!", SystemExit)
 
-	except:
-		utils.printf("Error while running threads")
+	except Exception as error:
+		utils.die("Error while running", error)
 
 	finally:
 		try: